fix: stream GTFS uploads directly to Vercel Blob to bypass 4MB API limit
- Replace buffered multipart upload (hit Next.js body limit) with client-side
  direct upload via @vercel/blob/client upload()
- Add /api/gtfs/upload-token for auth-gated Blob token generation
- Add /api/gtfs/register to create DB row + trigger processing after upload
- Update /api/gtfs/process to accept blobUrl directly instead of listing blobs
- Fix auth redirect loop: /gtfs no longer redirects server-side; shows inline
  sign-in gate when unauthenticated (passes authenticated prop to GtfsClient)
