+++ client/app/auth/signin/page.tsx
@@ -33,11 +33,11 @@ export default async function SignInPage({
 }: {
   searchParams: Promise<{ callbackUrl?: string }>;
 }) {
-  const session = await auth();
-  if (session) redirect("/");
-
   const { callbackUrl } = await searchParams;
   const callback = callbackUrl ?? "/";
+
+  const session = await auth();
+  if (session) redirect("/");
 
   return (
     <div className="flex min-h-screen flex-col items-center justify-center bg-gray-50 px-4">
